resolved CR comments for ODataBatchReader/Writer
