+++ src/Microsoft.OData.Core/MultipartMixed/ODataMultipartMixedBatchWriter.cs
@@ -69,14 +69,6 @@ namespace Microsoft.OData.Core.MultipartMixed
         }
 
         /// <summary>
-        /// Starts a new batch - implementation of the actual functionality.
-        /// </summary>
-        protected override void WriteStartBatchImplementation()
-        {
-            this.SetState(BatchWriterState.BatchStarted);
-        }
-
-        /// <summary>
         /// Ends a batch - implementation of the actual functionality.
         /// </summary>
         protected override void WriteEndBatchImplementation()
